plants_project/requirements.txt
pytest
pytest-django
django
pytest-cov
djangorestframework
